1.0

+++ Load.py
@@ -1,9 +1,16 @@
 import sys
 import pandas as pd
+import os.path
 
 def create():
-    with open("expenses.csv","w") as file:
+    file_exists=os.path.exists("expenses.csv")
+    if(file_exists):
         pass
+    else:
+        with open("expenses.csv","w") as file:
+            pass
+
+create()
